bot/settings.py
import os

from aiogram import Bot, Dispatcher
from aiogram.contrib.fsm_storage.memory import MemoryStorage
from aiogram.dispatcher.filters.state import StatesGroup, State
from dotenv import load_dotenv

load_dotenv()


class Config:
    BOT_TOKEN = os.getenv("BOT_TOKEN")
    bot = Bot(token=BOT_TOKEN)
    storage = MemoryStorage()
    dp = Dispatcher(bot, storage=storage)


config = Config()
bot = config.bot

dp = config.dp


class Form(StatesGroup):
    user_info = State()
    password = State()
    age = State()
    gender = State()
    phone = State()


form = Form()
