requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2023.2
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.8.2
